OBS-21 - Split required files out
