+++ include/plugin/Types.h
@@ -1,0 +1,27 @@
+#ifndef TYPES_H
+#define TYPES_H
+
+enum SourceCaptionerStatusEvent {
+    SOURCE_CAPTIONER_STATUS_EVENT_STOPPED,
+
+    SOURCE_CAPTIONER_STATUS_EVENT_STARTED_OK,
+    SOURCE_CAPTIONER_STATUS_EVENT_STARTED_ERROR,
+
+    SOURCE_CAPTIONER_STATUS_EVENT_NEW_SETTINGS_STOPPED,
+
+    SOURCE_CAPTIONER_STATUS_EVENT_AUDIO_CAPTURE_STATUS_CHANGE,
+};
+
+enum CaptionSourceMuteType {
+    CAPTION_SOURCE_MUTE_TYPE_FROM_OWN_SOURCE,
+    CAPTION_SOURCE_MUTE_TYPE_ALWAYS_CAPTION,
+    CAPTION_SOURCE_MUTE_TYPE_USE_OTHER_MUTE_SOURCE,
+};
+
+enum CapitalizationType {
+    CAPITALIZATION_NORMAL = 0,
+    CAPITALIZATION_ALL_CAPS = 1,
+    CAPITALIZATION_ALL_LOWERCASE = 2,
+};
+
+#endif //TYPES_H
