+++ src/window/MainWindow.cpp
@@ -1,4 +1,6 @@
-#include "MainWindow.h"
+#include <MainWindow.h>
+#include <ui_MainWindow.h>
+
 
 MainWindow::MainWindow(): QWidget(), ui(new Ui::MainWindow)
 {
